Tidy up, extract PlainValueInput
